Update fetch-entities.yml
Removed schedule until this can be fully tested and compared with the data from Huginn. Once fully tested, this schedule should be enabled, and the Huginn agent closed along with removing the grandtheatre-qc-ca-event-type-mapping graph, event type mapping file and related workflow (mapping event types is now included in this workflow so the separate workflow and the mapping files should be removed).